feat(trading-cycle): cycleSeedType 도입 + 사이클 자동 재등록 구현
- TradingCycle record에 cycleSeedType 필드 추가 (NONE/MAINTAIN/MAX)
- TradingCycleHistory.holdings BigDecimal → int 변환
- V47 마이그레이션: cycle_seed_type 컬럼 추가(테이블 재생성) + holdings INT 변환
- UpdateTradingCycleUseCase 신설, PUT /api/trading-cycles/{id} 엔드포인트 추가
- RegisterTradingCycleUseCase.Command에 cycleSeedType 추가
- MetaController: GET /api/meta/cycle-seed-types 추가, MetaBundle에 포함
- TradingService: holdings=0 시 MAINTAIN/MAX 자동 재등록 + 성공·실패 알림 발송
- TradingServiceTest: 재등록 시나리오 4개 추가(MAINTAIN/MAX 성공, 잔고부족, KIS오류)

+++ src/main/java/com/kista/domain/model/tradingcycle/TradingCycleHistory.java
@@ -10,6 +10,6 @@ public record TradingCycleHistory(
         UUID tradingCycleId,    // FK → trading_cycle.id (UUID 간접 참조)
         BigDecimal usdDeposit,  // 통합주문가능금액 (매매 공식 B 계산 기준)
         BigDecimal avgPrice,    // 평균 매입 단가 (보유수량 0이면 null)
-        BigDecimal holdings,    // 보유 수량
+        int holdings,    // 보유 수량
         Instant createdAt       // 생성 시각 (null이면 DB DEFAULT)
 ) {}
